podcast_crawler::infrastructure::config
Module utils

Configuration utility functions.

This module provides utility functions for configuration management including:

Environment variable parsing

String value handling

Test environment support

Example

use podcast_crawler::infrastructure::config::utils::{parse_env, get_env_string};

fn load_config() -> Result<(), Box<dyn std::error::Error>> {
    // Parse numeric value
    let port: u16 = parse_env("PORT")?;

    // Get string value
    let host = get_env_string("HOST")?;

    println!("Server will listen on {}:{}", host, port);
    Ok(())
}

Functions

get_env_string

Gets a string value from an environment variable

get_env_string_with_test

Gets a string value from an environment variable with test support

parse_env

Parses an environment variable into a specified type