podcast_crawler::infrastructure::config::utils
Function get_env_string

pub fn get_env_string(env_var: &str) -> AppResult<String>

Gets a string value from an environment variable

This function retrieves a string value from the specified environment variable.

Arguments

env_var - The name of the environment variable to read

Returns

Returns a Result containing the string value if successful, or an error if the environment variable is not found.

Example

use podcast_crawler::infrastructure::config::utils::get_env_string;

let host = get_env_string("HOST")?;
assert!(!host.is_empty());